feat: LegacyPostCard 및 PostDetail 컴포넌트 추가, PostCard 리팩토링 및 API 연동 개선
