+++ fundamentals/today-i-learned/src/pages/newHome/components/FilterSection.tsx
@@ -1,4 +1,3 @@
-import * as React from "react";
 import type { FilterSectionProps, SortOption } from "../utils/types";
 
 interface SortTab {
